added basic test
