--- src/Gooi.Azure/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace Skinder.Gooi.Azure;
-public class Class1
-{
-
-}
