--- src/Gooi.Core/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace Skinder.Gooi.Core;
-public class Class1
-{
-
-}
