+++ tests/Gooi.Tests/ConnectionProperties.cs
@@ -1,0 +1,8 @@
+using Skinder.Gooi.Core.Options;
+
+namespace Skinder.Gooi.Tests;
+
+public class ConnectionProperties: IConnectionProperties
+{
+  public string ConnectionString { get; set; }
+}
