--- tests/Gooi.Tests/UnitTest1.cs
@@ -1,18 +1,0 @@
-//using Gooi.Domain.Features.AzureMessage;
-using Moq;
-
-namespace Skinder.Gooi.Tests;
-
-public class UnitTest1
-{
-    [Fact]
-    public void Test1()
-    {
-        // var writer = new Mock<IOutputWriter>();
-        // var messageHandler = new MessageHandler();
-        //
-        // var result = messageHandler.Gooi();
-        //
-        // Assert.AreEqual("", result);
-    }
-}
